requirements.txt
google-generativeai==0.8.1
python-docx==0.8.11
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0